host/requirements.txt
pyserial==3.5
streamlit==1.30.0
pandas==2.0
